fix(engine): isBeingConstructed flag is not get out of sync
